+++ noethys/Ctrl/CTRL_TaskBarIcon.py
@@ -84,7 +84,7 @@ class CustomTaskBarIcon():
 
         # Rond rouge
         hauteurRond = hauteurTexte + padding * 2
-        largeurRond = largeurTexte + padding * 2 + hauteurRond/2.0
+        largeurRond = largeurTexte + padding * 2 + int(hauteurRond/2.0)
         if largeurRond < hauteurRond :
             largeurRond = hauteurRond
 
@@ -99,8 +99,8 @@ class CustomTaskBarIcon():
             dc.DrawRoundedRectangleRect(wx.Rect(xRond, yRond, largeurRond, hauteurRond), hauteurRond / 2.0)
 
         # Texte
-        xTexte = xRond + largeurRond / 2.0 - largeurTexte / 2.0
-        yTexte = yRond + hauteurRond / 2.0 - hauteurTexte / 2.0 - 1
+        xTexte = int(xRond + largeurRond / 2.0 - largeurTexte / 2.0)
+        yTexte = int(yRond + hauteurRond / 2.0 - hauteurTexte / 2.0 - 1)
         dc.DrawText(texte, xTexte, yTexte)
 
         mdc.SelectObject(wx.NullBitmap)
